requirements.txt
# Core LangGraph e LangChain
langchain>=0.1.0
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-community>=0.0.20

# Modelos de IA
langchain-google-genai>=1.0.0
langchain-openai>=0.0.8

# Ferramentas de busca
tavily-python>=0.3.0

# Interface Web
streamlit>=1.30.0

# Utilitários
python-dotenv>=1.0.0
pydantic>=2.0.0

# Gmail Integration (para projeto 3)
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Processamento e validação
requests>=2.31.0
typing-extensions>=4.8.0

# Desenvolvimento e debugging
rich>=13.0.0
pandas>=2.0.0 